requirements-screenshots.txt
# Requirements for screenshot generation
playwright>=1.45.0
Pillow>=10.0.0
